[wip] desktop frame-manager
I introduce the desktop frame manager, with the goal that the
application frame top level sheet will be adopted from desktop pane
and not from the graft.

For the moment the desktop frame manager is defined and it is the
default but its beahviur  is equal to stack frame manager i.e. the
top-level-sheet are adopted from the graft.