frame-manager: abanoned the idea to use several frame-manager
Using several frame-manager is problematic, changing frame-manager is
problematic, (th default is that when the frame is disowned all the x
server (windows) object are discarded).

Therefore from this commit I experiment with nly two frame manager,
one for the *window-manager* app and another one for all the other
applications.